# .specify/specs/002-account-auth-onboarding/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: News Pulse — 계정·인증·온보딩·인가 API
  version: 1.0.0
  description: |
    spec 002 "계정·인증·온보딩·인가" API Contract.
    모든 시각은 UTC ISO-8601. 액세스 토큰 TTL 1h, 리프레시 토큰 TTL 30d.
    `/api/v1/admin/**` — ADMIN 역할 전용.
    `/api/v1/**` (나머지) — 인증 필수 (이메일 인증 완료 포함).
    공개 화이트리스트: 인증·비밀번호 재설정·헬스체크 엔드포인트. 이메일 인증 엔드포인트는 인증 필수(토큰 필요, emailVerified=false 허용).
servers:
  - url: http://localhost:8080
    description: 로컬 개발
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      description: RFC 7807 Problem Details 기반 에러 응답
      properties:
        code:
          type: string
          example: INVALID_PASSWORD
        message:
          type: string
          example: 비밀번호는 영문·숫자를 포함한 8자 이상이어야 합니다
        timestamp:
          type: string
          format: date-time
          example: '2026-06-11T10:00:00Z'
      required: [code, message, timestamp]
...
    AccountSummary:
      type: object
      description: 현재 인증된 계정 요약 정보
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
          nullable: true
          description: 소셜 전용 계정은 null 가능
        role:
          type: string
          enum: [USER, ADMIN]
        emailVerified:
          type: boolean
        onboardingCompleted:
          type: boolean
        signupType:
          type: string
          enum: [EMAIL, SOCIAL]
        createdAt:
          type: string
          format: date-time
        requiresReConsent:
          type: boolean
          description: 현재 활성 필수 약관 중 미동의 버전 존재 여부. true이면 클라이언트가 재동의 UI를 표시해야 함
      required: [id, role, emailVerified, onboardingCompleted, signupType, createdAt, requiresReConsent]

    # ── Signup / Login ───────────────────────────────────────────────────────
...
    PasswordResetRequestBody:
      type: object
      description: 비밀번호 재설정 코드 발급 요청
      properties:
        email:
          type: string
          format: email
          description: |
            가입된 이메일로 코드를 발송함. 미가입 이메일도 동일 응답(계정 열거 방지, FR-026).
      required: [email]
    PasswordResetVerifyRequest:
      type: object
      description: 비밀번호 재설정 코드 검증
      properties:
        email:
          type: string
          format: email
        code:
          type: string
          description: 이메일로 받은 6자리 숫자 코드
          pattern: '^\d{6}$'
          example: '123456'
      required: [email, code]
    PasswordResetVerifyResponse:
      type: object
      properties:
        resetToken:
          type: string
          description: 비밀번호 변경 허가 토큰 (단일 사용, TTL 10분)
      required: [resetToken]
    PasswordResetConfirmRequest:
      type: object
      properties:
        resetToken:
          type: string
          description: PasswordResetVerifyResponse에서 받은 변경 허가 토큰
        newPassword:
          type: string
          description: 새 비밀번호 (영문·숫자 각 1자 이상, 8자 이상)
          minLength: 8
      required: [resetToken, newPassword]

    # ── Email Verification ────────────────────────────────────────────────────
...
    UserProfile:
      type: object
      properties:
        nickname:
          type: string
          maxLength: 50
          nullable: true
        ageGroup:
          type: string
          enum: [TEENS, TWENTIES, THIRTIES, FORTIES, FIFTIES_PLUS]
          nullable: true
        occupation:
          type: string
          maxLength: 100
          nullable: true
    UserInterests:
      type: object
      properties:
        categories:
          type: array
          description: 관심 카테고리 목록. 최소 3개 유지 필수.
          items:
            type: string
            enum: [POLITICS, ECONOMY, TECH, CULTURE, SPORTS, WORLD, SCIENCE, HEALTH, ENTERTAINMENT, LIFESTYLE]
          minItems: 3
      required: [categories]
    FollowKeywordsRequest:
      type: object
      properties:
        keywords:
          type: array
          items:
            $ref: '#/components/schemas/FollowKeywordItem'
      required: [keywords]
    FollowKeywordItem:
      type: object
      properties:
        keyword:
          type: string
          maxLength: 100
        type:
          type: string
          enum: [COMPANY, THEME, PERSON]
      required: [keyword, type]
    ReadingPreference:
      type: object
      properties:
        summaryDepth:
          type: string
          enum: [BRIEF, BALANCED, DEEP]
          description: "요약 깊이 (기본값: BALANCED)"
        consumeMode:
          type: string
          enum: [READ, LISTEN, BOTH]
          description: "소비 방식 (기본값: READ)"
      required: [summaryDepth, consumeMode]
    BriefingSettings:
      type: object
      properties:
        briefingTime:
          type: string
          pattern: '^([01]\d|2[0-3]):[0-5]\d$'
          description: 브리핑 시간 HH:MM (로컬 기준)
          example: '08:00'
        timezoneOffset:
          type: integer
          description: UTC 오프셋(분). KST=540, JST=540, UTC=0
          example: 540
        voiceEnabled:
          type: boolean
        pushAgreed:
          type: boolean
          description: 푸시 알림 동의. true 설정 시 동의 시각이 기록됨.
      required: [briefingTime, timezoneOffset, voiceEnabled, pushAgreed]
...
paths:
...
  /api/v1/auth/password-reset/request:
    post:
      tags: [Auth - Password Reset]
      summary: 비밀번호 재설정 코드 발급 (1단계)
      description: |
        등록·미등록·소셜 전용 계정 이메일 모두 동일하게 202를 반환한다 (계정 열거 방지, FR-026).
        이메일 서비스 장애 시 503 반환, 한도 미차감 (FR-011).
        시간당 5회 초과 시 429 반환 (FR-011).
        소셜 전용 계정 이메일로 요청 시 HTTP 응답은 동일 202이지만, 내부적으로 "비밀번호 재설정 불가 — 소셜 로그인 이용" 안내 이메일을 발송한다 (코드는 생성하지 않음).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequestBody'
      responses:
        '202':
          description: 처리 완료 (등록·미등록·소셜 전용 계정 모두 동일 202 — 계정 존재 여부 누설 없음)
        '429':
          description: 재전송 횟수 초과 (시간당 5회)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: 이메일 서비스 장애
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/password-reset/verify:
    post:
      tags: [Auth - Password Reset]
      summary: 비밀번호 재설정 코드 검증 (2단계)
      description: |
        이메일로 받은 6자리 코드를 검증한다.
        성공 시 단일 사용 변경 허가 토큰(TTL 10분)을 반환한다.
        코드 만료 시 410, 오입력 401(시도 횟수 증가), 5회 초과 시 코드 무효화.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetVerifyRequest'
      responses:
        '200':
          description: 코드 검증 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordResetVerifyResponse'
        '401':
          description: 잘못된 코드
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '410':
          description: 코드 만료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/password-reset/confirm:
    post:
      tags: [Auth - Password Reset]
      summary: 새 비밀번호 설정 (3단계)
      description: |
        변경 허가 토큰과 새 비밀번호로 비밀번호를 재설정한다.
        성공 시 해당 계정의 모든 리프레시 토큰을 즉시 무효화한다 (FR-025).
        변경 허가 토큰은 단일 사용 — 재사용 시 401.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirmRequest'
      responses:
        '204':
          description: 비밀번호 변경 완료 (본문 없음)
        '401':
          description: 변경 허가 토큰 무효·만료·재사용 시도
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 비밀번호 정책 미충족
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ═══════════════════════════════════════════════════════════════════════════
  # EMAIL VERIFICATION — Authenticated (email-unverified accounts allowed)
  # ═══════════════════════════════════════════════════════════════════════════
...
  /api/v1/me:
    get:
      tags: [Me]
      summary: 내 계정 요약 조회
      description: |
        현재 인증된 계정의 기본 정보를 반환한다.
        이메일 인증이 완료된 계정만 접근 가능 (emailVerified=true 필수).
        미인증 계정은 403(code=EMAIL_NOT_VERIFIED) 반환.
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountSummary'
        '401':
          description: 미인증 (토큰 없음 또는 유효하지 않음)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 이메일 미인증 (code=EMAIL_NOT_VERIFIED). 이메일 인증 완료 후 재로그인하여 새 토큰 발급 필요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/me/profile:
    get:
      tags: [Me]
      summary: 프로필 조회
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: 미인증
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Me]
      summary: 프로필 수정
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfile'
      responses:
        '200':
          description: 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: 미인증
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 유효성 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/me/interests:
    get:
      tags: [Me]
      summary: 관심 카테고리 조회
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInterests'
        '401':
          description: 미인증
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Me]
      summary: 관심 카테고리 수정 (최소 3개 유지)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInterests'
      responses:
        '200':
          description: 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInterests'
        '401':
          description: 미인증
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 카테고리 3개 미만
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/me/keywords:
    get:
      tags: [Me]
      summary: 팔로우 키워드 조회
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FollowKeywordsRequest'
        '401':
          description: 미인증
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Me]
      summary: 팔로우 키워드 교체 (기존 목록 전체 교체)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FollowKeywordsRequest'
      responses:
        '200':
          description: 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FollowKeywordsRequest'
        '401':
          description: 미인증
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/me/reading-preference:
    get:
      tags: [Me]
      summary: 읽는 방식 설정 조회
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadingPreference'
        '401':
          description: 미인증
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Me]
      summary: 읽는 방식 설정 수정
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReadingPreference'
      responses:
        '200':
          description: 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadingPreference'
        '401':
          description: 미인증
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/me/briefing-settings:
    get:
      tags: [Me]
      summary: 브리핑 설정 조회
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BriefingSettings'
        '401':
          description: 미인증
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Me]
      summary: 브리핑 설정 수정
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BriefingSettings'
      responses:
        '200':
          description: 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BriefingSettings'
        '401':
          description: 미인증
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...