# src/main/java/com/newscurator/domain/UserInterests.java
package com.newscurator.domain;

import jakarta.persistence.*;
import java.util.UUID;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;

@Entity
@Table(name = "user_interests",
        uniqueConstraints = @UniqueConstraint(columnNames = {"account_id", "category"}))
@Getter
@NoArgsConstructor
public class UserInterests {

    @Id
    @GeneratedValue(strategy = GenerationType.UUID)
    private UUID id;

    @ManyToOne(fetch = FetchType.LAZY)
    @JoinColumn(name = "account_id", nullable = false)
    private Account account;

    @Column(nullable = false, length = 50)
    private String category;

    @Builder
    public UserInterests(Account account, String category) {
        this.account = account;
        this.category = category;
    }
}